+++ Sources/VComponents/Components/Buttons/Loading Stretched Button/VLoadingStretchedButtonLabel.swift
@@ -1,5 +1,5 @@
 //
-//  VPrimaryButtonLabel.swift
+//  VLoadingStretchedButtonLabel.swift
 //  VComponents
 //
 //  Created by Vakhtang Kontridze on 2/26/22.
@@ -7,8 +7,8 @@
 
 import SwiftUI
 
-// MARK: - V Primary Button Label
+// MARK: - V Loading Stretched Button Label
 @available(macOS, unavailable)
 @available(tvOS, unavailable)
 @available(watchOS, unavailable)
-typealias VPrimaryButtonLabel = GenericContent_TitleIconTitleContent
+typealias VLoadingStretchedButtonLabel = GenericContent_TitleIconTitleContent