+++ Sources/VComponents/Components/Buttons/Loading Stretched Button/VLoadingStretchedButtonState.swift
@@ -1,5 +1,5 @@
 //
-//  VPrimaryButtonState.swift
+//  VLoadingStretchedButtonState.swift
 //  VComponents
 //
 //  Created by Vakhtang Kontridze on 19.12.20.
@@ -8,8 +8,8 @@
 import SwiftUI
 import VCore
 
-// MARK: - V Primary Button Internal State
+// MARK: - V Loading Stretched Button Internal State
 @available(macOS, unavailable)
 @available(tvOS, unavailable)
 @available(watchOS, unavailable)
-typealias VPrimaryButtonInternalState = GenericState_EnabledPressedLoadingDisabled
+typealias VLoadingStretchedButtonInternalState = GenericState_EnabledPressedLoadingDisabled